fix: 38 audit findings across two rounds (memory safety, correctness, hardening)
Round 1 — 20 fixes (prior audit + fresh findings):

HIGH (5):
- reader.c: depth-guard bypass for #[/#{/#comprehensions
- adr.h, adfmt.c: trailing-backslash OOB read (4 sites)
- debugger.h: macro closure env loss in expandmacro
- resp.c: epoch registration exhaustion UAF
- resp.c: FD_SET stack buffer overflow + connection cap

MEDIUM (10):
- compiler_impl.h: compile_expr depth guard
- alcove.c: savedbcmd refcount leak + wrong return
- alcove.c: updatebang refcount leak + NULL return
- resp.c: cmd_lpop_rpop NULL deref on malloc failure
- persist.h: load_callable truncated-read checks
- vector.h: vec-push! defers end++ after store
- builtins_stdlib.h: trycmd in_tail_position save/clear/restore
- adfmt.c: tab/space mix detection
- resp.c: connection cap (RESP_MAX_CLIENTS)
- ffi.h: FFI callback lifetime documentation

LOW (5):
- resp.c: rbuf/wbuf OOM drops client not server
- reader.c: hex literal FIX_FITS (wasm32)
- ffi.h: per-field bounds assertion in ffi-pack
- dict.h: rehash overflow guard

Round 2 — 18 fixes (GC/weak/watch, containers, numeric, pp, env/dict):

HIGH (2):
- print.h: circular list hang (element-count cap)
- watch.h: watch_notify UAF (save c->next before EVAL)

MEDIUM (8):
- numeric.h: INT64_MIN negation UB, decimal div overflow x2
- json.h: FIX_FITS for numbers + fix js_putn missing return
- msgpack.h: uint64 silent wrap
- dict.h: checked strdup (dict_strdup)
- vector.h: watch_notify on vec-push!/pop!
- persist.h: load_callable params leak on body failure
- compiler_impl.h: stack_guard_exhausted replaces compile_depth

LOW (8):
- json.h: reject \u0000
- hamt.h: reject NaN keys
- pp.h: depth limit
- print.h: string escaping
- numeric.h: decimal cmp string fallback
- msgpack.h: array/map length caps
- dict.h: bernstein_hash size_t len
- watch.h: safe iteration past newly-added watchers

All builds pass. Tests: 2519+2625 passed, 0 failed. fmt-check clean.

+++ alcove.h
@@ -928,10 +928,10 @@ exp_t *set_return_point(env_t *env);
 dict_t *create_dict();
 int destroy_dict(dict_t *d);
 int dump_dict(dict_t *d, FILE *stream);
-unsigned int bernstein_hash(unsigned char *key, int len);
+unsigned int bernstein_hash(unsigned char *key, size_t len);
 unsigned int
 bernstein_hash_z(const char *key); /* fused strlen+hash, C-strings */
-unsigned int bernstein_uhash(unsigned char *key, int len);
+unsigned int bernstein_uhash(unsigned char *key, size_t len);
 keyval_t *set_get_keyval_dict(dict_t *d, char *key, exp_t *val);
 exp_t *set_keyval_dict_timestamp(dict_t *d, char *key, int64_t timestamp);
 int64_t get_keyval_dict_timestamp(dict_t *d, char *key);
